Add support for Windows pipes
