+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/radareorg/r2pipe-go/v3
 
-go 1.24
+go 1.24.0
+
+require golang.org/x/sys v0.40.0 // indirect
